Functions, the basics.
Learning the basics of functions. Creating, calling and
returning values.

Example function adds two numbers together, stores the
result, and returns the value.

Questions that come to mind:
While using a constant when applicable might make code easier
to understand when reading it, is there any other advantage?
My instinct would be to assign all values via 'let' instead,
and if it never changes then so be it.